Removing PowerConsoleCommands.cs because it's loading OleMenuCommandService( which only needed for PMC window load) at app start which is not good.
