+++ src/NuGet.Clients/NuGet.Console/NuGet.Console.csproj
@@ -57,7 +57,6 @@
     <Compile Include="OutputConsole\OutputConsoleProvider.cs" />
     <Compile Include="OutputConsole\BaseOutputConsole.cs" />
     <Compile Include="OutputConsole\SharedOutputConsole.cs" />
-    <Compile Include="PowerConsoleCommands.cs" />
     <Compile Include="ScriptExecutor.cs" />
     <Compile Include="SimpleExpansion.cs" />
     <Compile Include="Token.cs" />
