+++ src/NuGet.Clients/NuGet.Tools/NuGetPackage.cs
@@ -148,8 +148,6 @@ namespace NuGetVSExtension
             // Add our command handlers for menu (commands must exist in the .vsct file)
             await AddMenuCommandHandlersAsync();
 
-            await PowerConsoleCommands.Instance.InitializeAsync(this);
-
             // This instantiates a decoupled ICommand instance responsible to locate and display output pane by a UI control
             UI.Commands.ShowErrorsCommand = new ShowErrorsCommand(this);
